fixed small issues

+++ app/src/main/java/de/dali/thesisfingerprint2019/data/local/entity/FingerPrintEntity.kt
@@ -39,7 +39,7 @@ data class FingerPrintEntity(
 
     var resolution: String? = null,
 
-    var correctionDegree: Float? = null,
+    var correctionDegree: Double? = null,
 
     var imageList: List<String>? = null
